scanvi, scvi, autozi, linearscvi training
