+++ scvi/model/_scanvi.py
@@ -6,13 +6,13 @@ import pandas as pd
 import torch
 from anndata import AnnData
 from pandas.api.types import CategoricalDtype
-from sklearn.model_selection._split import _validate_shuffle_split
 
-from scvi import _CONSTANTS, settings
+from scvi import _CONSTANTS  # , settings
 from scvi._compat import Literal
+from scvi.compose._training import SemiSupervisedDataSplitter, TrainRunner
 from scvi.data._anndata import _make_obs_column_categorical
 from scvi.dataloaders import AnnDataLoader, SemiSupervisedDataLoader
-from scvi.lightning import SemiSupervisedTrainingPlan, Trainer
+from scvi.lightning import SemiSupervisedTrainingPlan
 from scvi.lightning._callbacks import SubSampleLabels
 from scvi.model._scvi import SCVI
 from scvi.modules import SCANVAE
@@ -238,98 +238,6 @@ class SCANVI(RNASeqMixin, VAEMixin, ArchesMixin, BaseModelClass):
     def _data_loader_cls(self):
         return AnnDataLoader
 
-    @property
-    def history(self):
-        """Returns computed metrics during training."""
-        return self._trainer.logger.history
-
-    def train(
-        self,
-        max_epochs: Optional[int] = None,
-        n_samples_per_label: Optional[float] = None,
-        check_val_every_n_epoch: Optional[int] = None,
-        train_size: float = 0.9,
-        validation_size: Optional[float] = None,
-        batch_size: int = 128,
-        use_gpu: Optional[bool] = None,
-        plan_kwargs: Optional[dict] = None,
-        **kwargs,
-    ):
-        """
-        Train the model.
-
-        Parameters
-        ----------
-        max_epochs
-            Number of passes through the dataset for semisupervised training.
-        n_samples_per_label
-            Number of subsamples for each label class to sample per epoch. By default, there
-            is no label subsampling.
-        check_val_every_n_epoch
-            Frequency with which metrics are computed on the data for validation set for both
-            the unsupervised and semisupervised trainers. If you'd like a different frequency for
-            the semisupervised trainer, set check_val_every_n_epoch in semisupervised_train_kwargs.
-        train_size
-            Size of training set in the range [0.0, 1.0].
-        validation_size
-            Size of the test set. If `None`, defaults to 1 - `train_size`. If
-            `train_size + validation_size < 1`, the remaining cells belong to a test set.
-        batch_size
-            Minibatch size to use during training.
-        use_gpu
-            If `True`, use the GPU if available. Will override the use_gpu option when initializing model
-        plan_kwargs
-            Keyword args for :class:`~scvi.lightning.SemiSupervisedTrainingPlan`. Keyword arguments passed to
-            `train()` will overwrite values present in `plan_kwargs`, when appropriate.
-        **kwargs
-            Other keyword args for :class:`~scvi.lightning.Trainer`.
-        """
-        if max_epochs is None:
-            n_cells = self.adata.n_obs
-            max_epochs = np.min([round((20000 / n_cells) * 400), 400])
-
-        logger.info("Training for {} epochs.".format(max_epochs))
-
-        use_gpu = use_gpu if use_gpu is not None else self.use_gpu
-        gpus = 1 if use_gpu else None
-        pin_memory = (
-            True if (settings.dl_pin_memory_gpu_training and use_gpu) else False
-        )
-        train_dl, val_dl, test_dl = self._train_test_val_split(
-            self.adata,
-            unlabeled_category=self.unlabeled_category_,
-            train_size=train_size,
-            validation_size=validation_size,
-            n_samples_per_label=n_samples_per_label,
-            pin_memory=pin_memory,
-            batch_size=batch_size,
-        )
-
-        self.train_indices_ = train_dl.indices
-        self.validation_indices_ = val_dl.indices
-        self.test_indices_ = test_dl.indices
-
-        plan_kwargs = {} if plan_kwargs is None else plan_kwargs
-        self._task = SemiSupervisedTrainingPlan(self.module, **plan_kwargs)
-
-        # if we have labeled cells, we want to subsample labels each epoch
-        sampler_callback = (
-            [SubSampleLabels()] if len(self._labeled_indices) != 0 else []
-        )
-        self._trainer = Trainer(
-            max_epochs=max_epochs,
-            gpus=gpus,
-            callbacks=sampler_callback,
-            check_val_every_n_epoch=check_val_every_n_epoch,
-            **kwargs,
-        )
-        if len(self.validation_indices_) != 0:
-            self._trainer.fit(self._task, train_dl, val_dl)
-        else:
-            self._trainer.fit(self._task, train_dl)
-        self.module.eval()
-        self.is_trained_ = True
-
     def predict(
         self,
         adata: Optional[AnnData] = None,
@@ -386,139 +294,90 @@ class SCANVI(RNASeqMixin, VAEMixin, ArchesMixin, BaseModelClass):
             )
             return y_pred
 
-    def _train_test_val_split(
+    def train(
         self,
-        adata: AnnData,
-        unlabeled_category,
+        max_epochs: Optional[int] = None,
+        n_samples_per_label: Optional[float] = None,
+        check_val_every_n_epoch: Optional[int] = None,
         train_size: float = 0.9,
         validation_size: Optional[float] = None,
-        n_samples_per_label: Optional[int] = None,
-        **kwargs,
+        batch_size: int = 128,
+        use_gpu: Optional[bool] = None,
+        plan_kwargs: Optional[dict] = None,
+        **trainer_kwargs,
     ):
         """
-        Creates data loaders ``train_set``, ``validation_set``, ``test_set``.
-
-        If ``train_size + validation_set < 1`` then ``test_set`` is non-empty.
-        The ratio between labeled and unlabeled data in adata will be preserved
-        in the train/test/val sets.
+        Train the model.
 
         Parameters
         ----------
-        adata
-            AnnData to split into train/test/val sets
-        unlabeled_category
-            Category to treat as unlabeled
-        train_size
-            float, or None (default is 0.9)
-        validation_size
-            float, or None (default is None)
+        max_epochs
+            Number of passes through the dataset for semisupervised training.
         n_samples_per_label
-            Number of subsamples for each label class to sample per epoch
-        **kwargs
-            Keyword args for `_make_data_loader()`
+            Number of subsamples for each label class to sample per epoch. By default, there
+            is no label subsampling.
+        check_val_every_n_epoch
+            Frequency with which metrics are computed on the data for validation set for both
+            the unsupervised and semisupervised trainers. If you'd like a different frequency for
+            the semisupervised trainer, set check_val_every_n_epoch in semisupervised_train_kwargs.
+        train_size
+            Size of training set in the range [0.0, 1.0].
+        validation_size
+            Size of the test set. If `None`, defaults to 1 - `train_size`. If
+            `train_size + validation_size < 1`, the remaining cells belong to a test set.
+        batch_size
+            Minibatch size to use during training.
+        use_gpu
+            If `True`, use the GPU if available. Will override the use_gpu option when initializing model
+        plan_kwargs
+            Keyword args for :class:`~scvi.lightning.SemiSupervisedTrainingPlan`. Keyword arguments passed to
+            `train()` will overwrite values present in `plan_kwargs`, when appropriate.
+        **trainer_kwargs
+            Other keyword args for :class:`~scvi.lightning.Trainer`.
         """
-        train_size = float(train_size)
-        if train_size > 1.0 or train_size <= 0.0:
-            raise ValueError(
-                "train_size needs to be greater than 0 and less than or equal to 1"
-            )
 
-        n_labeled_idx = len(self._labeled_indices)
-        n_unlabeled_idx = len(self._unlabeled_indices)
+        if max_epochs is None:
+            n_cells = self.adata.n_obs
+            max_epochs = np.min([round((20000 / n_cells) * 400), 400])
 
-        def get_train_val_split(n_samples, test_size, train_size):
-            try:
-                n_train, n_val = _validate_shuffle_split(
-                    n_samples, test_size, train_size
-                )
-            except ValueError:
-                if train_size != 1.0 and n_samples != 1:
-                    raise ValueError(
-                        "Choice of train_size={} and validation_size={} not understood".format(
-                            train_size, test_size
-                        )
-                    )
-                n_train, n_val = n_samples, 0
-            return n_train, n_val
+        logger.info("Training for {} epochs.".format(max_epochs))
 
-        if n_labeled_idx != 0:
-            n_labeled_train, n_labeled_val = get_train_val_split(
-                n_labeled_idx, validation_size, train_size
-            )
-            labeled_permutation = np.random.choice(
-                self._labeled_indices, len(self._labeled_indices), replace=False
-            )
-            labeled_idx_val = labeled_permutation[:n_labeled_val]
-            labeled_idx_train = labeled_permutation[
-                n_labeled_val : (n_labeled_val + n_labeled_train)
-            ]
-            labeled_idx_test = labeled_permutation[(n_labeled_val + n_labeled_train) :]
-        else:
-            labeled_idx_test = []
-            labeled_idx_train = []
-            labeled_idx_val = []
+        plan_kwargs = {} if plan_kwargs is None else plan_kwargs
 
-        if n_unlabeled_idx != 0:
-            n_unlabeled_train, n_unlabeled_val = get_train_val_split(
-                n_unlabeled_idx, validation_size, train_size
-            )
-            unlabeled_permutation = np.random.choice(
-                self._unlabeled_indices, len(self._unlabeled_indices)
-            )
-            unlabeled_idx_val = unlabeled_permutation[:n_unlabeled_val]
-            unlabeled_idx_train = unlabeled_permutation[
-                n_unlabeled_val : (n_unlabeled_val + n_unlabeled_train)
-            ]
-            unlabeled_idx_test = unlabeled_permutation[
-                (n_unlabeled_val + n_unlabeled_train) :
-            ]
-        else:
-            unlabeled_idx_train = []
-            unlabeled_idx_val = []
-            unlabeled_idx_test = []
+        use_gpu = use_gpu if use_gpu is not None else self.use_gpu
+        gpus = 1 if use_gpu else None
+        # pin_memory = (
+        #     True if (settings.dl_pin_memory_gpu_training and use_gpu) else False
+        # )
 
-        indices_train = np.concatenate((labeled_idx_train, unlabeled_idx_train))
-        indices_val = np.concatenate((labeled_idx_val, unlabeled_idx_val))
-        indices_test = np.concatenate((labeled_idx_test, unlabeled_idx_test))
-
-        indices_train = indices_train.astype(int)
-        indices_val = indices_val.astype(int)
-        indices_test = indices_test.astype(int)
-
-        if len(self._labeled_indices) != 0:
-            data_loader_class = SemiSupervisedDataLoader
-            dl_kwargs = {
-                "unlabeled_category": unlabeled_category,
-                "n_samples_per_label": n_samples_per_label,
-            }
-        else:
-            data_loader_class = AnnDataLoader
-            dl_kwargs = {}
-        dl_kwargs.update(kwargs)
-
-        scanvi_train_dl = self._make_data_loader(
-            adata,
-            indices=indices_train,
-            shuffle=True,
-            data_loader_class=data_loader_class,
-            drop_last=3,
-            **dl_kwargs,
-        )
-        scanvi_val_dl = self._make_data_loader(
-            adata,
-            indices=indices_val,
-            shuffle=True,
-            data_loader_class=data_loader_class,
-            drop_last=3,
-            **dl_kwargs,
-        )
-        scanvi_test_dl = self._make_data_loader(
-            adata,
-            indices=indices_test,
-            shuffle=True,
-            data_loader_class=data_loader_class,
-            drop_last=3,
-            **dl_kwargs,
+        # if we have labeled cells, we want to subsample labels each epoch
+        sampler_callback = (
+            [SubSampleLabels()] if len(self._labeled_indices) != 0 else []
         )
 
-        return scanvi_train_dl, scanvi_val_dl, scanvi_test_dl
+        data_splitter = SemiSupervisedDataSplitter(
+            adata=self.adata,
+            unlabeled_category=self.unlabeled_category_,
+            train_size=train_size,
+            validation_size=validation_size,
+            n_samples_per_label=n_samples_per_label,
+            batch_size=batch_size,
+        )
+        training_plan = SemiSupervisedTrainingPlan(
+            self.module, len(data_splitter.train_idx), **plan_kwargs
+        )
+        if "callbacks" in trainer_kwargs.keys():
+            trainer_kwargs["callbacks"].concatenate(sampler_callback)
+        else:
+            trainer_kwargs["callbacks"] = sampler_callback
+
+        runner = TrainRunner(
+            self,
+            training_plan=training_plan,
+            data_splitter=data_splitter,
+            max_epochs=max_epochs,
+            gpus=gpus,
+            check_val_every_n_epoch=check_val_every_n_epoch,
+            **trainer_kwargs,
+        )
+        return runner()
